+++ tests/plotting/plot_2d_test.py
@@ -334,7 +334,7 @@ def test_plot_2d_binned_data_datetime64():
 
 def test_plot_3d_binned_data_where_outer_dimension_has_no_event_coord():
     data = make_binned_data_array(ndim=2, masks=True)
-    data = sc.concatenate(data, data * sc.scalar(2.0), 'run')
+    data = sc.concat([data, data * sc.scalar(2.0)], 'run')
     plot_obj = sc.plot(data)
     plot_obj.widgets._controls['run']['slider'].value = 1
     plot_obj.close()
@@ -342,7 +342,7 @@ def test_plot_3d_binned_data_where_outer_dimension_has_no_event_coord():
 
 def test_plot_3d_binned_data_where_inner_dimension_has_no_event_coord():
     data = make_binned_data_array(ndim=2)
-    data = sc.concatenate(data, data * sc.scalar(2.0), 'run')
+    data = sc.concat([data, data * sc.scalar(2.0)], 'run')
     plot(sc.transpose(data, dims=['yy', 'xx', 'run']))